+++ src/init/final/Finalizer.ts
@@ -1,0 +1,6 @@
+import { AsyncHandler } from '../../util/handlers/AsyncHandler';
+
+/**
+ * Finalizer is used to indicate an AsyncHandler that performs finalization logic.
+ */
+export abstract class Finalizer extends AsyncHandler {}