+++ test/unit/util/locking/FileSystemResourceLocker.test.ts
@@ -1,6 +1,5 @@
 import { readdir } from 'fs-extra';
-import { InternalServerError } from '../../../../src/util/errors/InternalServerError';
-import { FileSystemResourceLocker } from '../../../../src/util/locking/FileSystemResourceLocker';
+import { InternalServerError, FileSystemResourceLocker } from '../../../../src';
 
 const lockFolder = './.internal/locks/';
 
@@ -10,6 +9,7 @@ describe('A FileSystemResourceLocker', (): void => {
 
   beforeEach(async(): Promise<void> => {
     locker = new FileSystemResourceLocker({ attemptSettings: { retryCount: 19, retryDelay: 100 }});
+    await locker.initialize();
   });
 
   afterEach(async(): Promise<void> => {
@@ -109,12 +109,19 @@ describe('A FileSystemResourceLocker', (): void => {
     await expect(locker.acquire(identifier)).rejects.toThrow(InternalServerError);
   });
 
-  it('clears the files in de lock directory after calling finalize.', async(): Promise<void> => {
+  it('clears the files in de lock directory upon calling initialize.', async(): Promise<void> => {
     await locker.acquire(identifier);
     await expect(readdir(lockFolder)).resolves.toHaveLength(1);
-    await locker.finalize();
-    await expect(readdir(lockFolder)).rejects.toThrow();
+    await locker.initialize();
+    await expect(readdir(lockFolder)).resolves.toHaveLength(0);
   });
+
+  it('stops proper-lock from throwing errors onCompromise after finalize was called.',
+    async(): Promise<void> => {
+      expect((): void => (locker as any).customOnCompromised(new Error('test'))).toThrow();
+      await locker.finalize();
+      expect((locker as any).customOnCompromised(new Error('test'))).toBeUndefined();
+    });
 
   it('can create a locker with default AttemptSettings.', async(): Promise<void> => {
     expect((): FileSystemResourceLocker => new FileSystemResourceLocker()).not.toThrow();